+++ R/aaa.R
@@ -4,7 +4,7 @@ globalVariables(
     ".", ".x", "bi_prob", "community", "comp", "component", "component_new",
     "count_1", "count_2", "current_id", "dens", "dev_png", "doublet_nns",
     "doublet_p", "doublet_p_adj", "frequency", "from", "g", "graph_projection",
-    "group", "hjust", "host_freq", "host_unmixed_freq", "hp", "hup",
+    "group", "hjust", "receiver_freq", "receiver_unmixed_freq", "hp", "hup",
     "id", "id_map", "in_gate", "index", "item", "join_count", "join_count_expected_mean",
     "label", "layout", "marker", "marker_1", "marker_2", "marker_x",
     "marker_y", "marker1", "marker2", "med_ref", "med_tgt", "mixture_component",