+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Supervised patch detection implemented in the `patch_detection` function. Added `identify_markers_for_patch_analysis` to identify markers for patch analysis.
 - `render_rotating_layout` function to create videos of rotating cells from a tibble containing layout coordinates. The function supports multiple video formats but uses GIF as default.
 - `subset` method for `CellGraph` class
 - Option to add marker count proportions to the proximity score table in `ProximityScores`